chore: move CE-U related functions to `ceu.py`
